Python/Backend_Project/relivator-web-store-backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
PyJWT==2.10.1
python-dotenv==1.2.1
Werkzeug==3.1.4
